requirements.txt
aiogram>=3.4,<4
openai>=2.26,<3
python-dotenv>=1.0,<2
httpx>=0.27,<1
